requirements.txt
nicegui==1.4.12
duckdb==0.9.2
pyyaml==6.0.1
